The builder needs to use the run() helper too.
